PersonalProjects/CompliAI-Pro/backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pyyaml==6.0.2
python-multipart==0.0.9
pdfminer.six==20231228
pydantic==2.8.2